Reorganized files. Refactored code. Added extra credit from assignment2.
Created controller and design folders. Added scores and settings views.
Added GameResult model.

+++ iMatch/Model/Card.h
@@ -13,7 +13,7 @@
 @property (nonatomic, getter = isFaceUp) BOOL faceUp;
 @property (nonatomic, getter = isPlayable) BOOL playable;
 
-- (int) match:(NSArray *)otherCards;
+- (NSInteger) match:(NSArray *)otherCards;
 - (NSAttributedString *) attributedDescription;
 
 @end
