+++ iMatch/Model/Card.m
@@ -22,9 +22,9 @@
 }
 
 
-- (int)match:(NSArray *)otherCards
+- (NSInteger)match:(NSArray *)otherCards
 {
-    int score = 0;
+    NSInteger score = 0;
     for (id obj in otherCards)
     {
         // instrospection
